Enable driver comms for all chips, and strengthen i2c error checks.
The problem with the drivers not communicating properly was that the address pin 0
was connected to the error LED and therefore pulled low. This set the wrong address on the driver.
Replacing the resistor with a 10K ohm one fixes the problem for now, but it needs a chip redesign.